Move StakeAction to the types crate
